Move telemetry initializer
